Atmel AT32U4 version of the A-10C i2c master
This is the latest version of the controller for the DCS A-10C I2C master controller implemented in the CMSC.  This is the last version of the code before it was moved to the ESP8266 project.  This project communicates over serial to a C# program running on a Windows PC.  This was deprecated only because I wanted to get rid of the C# program, but it was functional.  It worked well for a long time.